requirements.txt
numpy==1.16.0
pandas==1.1.0
matplotlib==3.0.0
torch==1.9.0
torchvision==0.1.9
pillow==7.0
glob2==0.4.1